requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]
passlib
python-multipart
pillow
python-dotenv
email-validator
pydantic[email]
bcrypt>=4.0.1
passlib>=1.7.4
python-dotenv
psycopg2
supabase